# Changed hashing bit rotation to use Integer.rotateRight