Related to https://issues.scala-lang.org/browse/SI-10049 and https://issues.scala-lang.org/browse/SI-10084

The benchmarks show the following (`ns/op`, lower is better)

`Before:`
```scala
Benchmark                                     (size)  Mode  Cnt      Score      Error  Units
s.c.immutable.VectorMapBenchmark.groupBy          10  avgt   20    616.164 ±    4.712  ns/op
s.c.immutable.VectorMapBenchmark.groupBy         100  avgt   20   2034.447 ±   14.495  ns/op
s.c.immutable.VectorMapBenchmark.groupBy        1000  avgt   20  14712.164 ±  119.983  ns/op
s.c.mutable.HashMapBenchmark.get                  10  avgt   20    679.046 ±    6.872  ns/op
s.c.mutable.HashMapBenchmark.get                 100  avgt   20   7242.097 ±   41.244  ns/op
s.c.mutable.HashMapBenchmark.get                1000  avgt   20  95342.919 ± 1521.328  ns/op
s.c.mutable.HashMapBenchmark.getOrElseUpdate      10  avgt   20    488.034 ±    4.554  ns/op
s.c.mutable.HashMapBenchmark.getOrElseUpdate     100  avgt   20   4883.123 ±   59.268  ns/op
s.c.mutable.HashMapBenchmark.getOrElseUpdate    1000  avgt   20  65174.034 ±  496.759  ns/op
s.c.mutable.HashMapBenchmark.put                  10  avgt   20    267.983 ±    1.797  ns/op
s.c.mutable.HashMapBenchmark.put                 100  avgt   20   5097.351 ±  104.538  ns/op
s.c.mutable.HashMapBenchmark.put                1000  avgt   20  78772.540 ±  543.935  ns/op
```
`After:`
```scala
Benchmark                                     (size)  Mode  Cnt      Score      Error  Units
s.c.immutable.VectorMapBenchmark.groupBy          10  avgt   20    600.785 ±    1.639  ns/op
s.c.immutable.VectorMapBenchmark.groupBy         100  avgt   20   1927.722 ±   10.543  ns/op
s.c.immutable.VectorMapBenchmark.groupBy        1000  avgt   20  14450.596 ±   86.490  ns/op
s.c.mutable.HashMapBenchmark.get                  10  avgt   20    654.584 ±    3.316  ns/op
s.c.mutable.HashMapBenchmark.get                 100  avgt   20   7212.925 ±   43.364  ns/op
s.c.mutable.HashMapBenchmark.get                1000  avgt   20  94023.279 ± 1618.070  ns/op
s.c.mutable.HashMapBenchmark.getOrElseUpdate      10  avgt   20    485.574 ±    3.803  ns/op
s.c.mutable.HashMapBenchmark.getOrElseUpdate     100  avgt   20   4920.506 ±   25.781  ns/op
s.c.mutable.HashMapBenchmark.getOrElseUpdate    1000  avgt   20  63444.007 ±  289.288  ns/op
s.c.mutable.HashMapBenchmark.put                  10  avgt   20    241.557 ±    1.778  ns/op
s.c.mutable.HashMapBenchmark.put                 100  avgt   20   5125.638 ±   30.796  ns/op
s.c.mutable.HashMapBenchmark.put                1000  avgt   20  77419.913 ±  548.382  ns/op
```