@paplorinc 2016-12-06
@Ichoran, @retronym, could you please review this :)?